requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
idna==2.8
lxml==4.5.0
multidict==4.7.4
pyquery==1.4.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
tzlocal==2.0.0
urllib3==1.25.8
yarl==1.4.2
